services/worker/requirements.txt
psycopg2-binary
pandas
numpy
redis>=5.0
rq>=1.15 